Split tests for Scala 2 and 3
