+++ airframe/src/test/scala/example/SealedTrait.scala
@@ -1,0 +1,11 @@
+package example
+
+/**
+  */
+object SealedTrait {
+  sealed trait Adt
+  object Adt {
+    object Foo extends Adt
+    object Bar extends Adt
+  }
+}
