+++ airframe/src/test/scala/wvlet/airframe/di/DefaultValueTest.scala
@@ -1,5 +1,6 @@
 package wvlet.airframe.di
 
+import wvlet.airframe.Design
 import wvlet.airspec.AirSpec
 
 /**
