+++ airframe/src/test/scala/wvlet/airframe/di/ProviderTest.scala
@@ -13,6 +13,7 @@
  */
 package wvlet.airframe.di
 
+import wvlet.airframe.Design
 import wvlet.airspec.AirSpec
 import wvlet.log.LogSupport
