+++ airframe/src/test/scala/wvlet/airframe/di/SessionBuilderTest.scala
@@ -13,7 +13,8 @@
  */
 package wvlet.airframe.di
 
-import wvlet.airframe.di.lifecycle.{LifeCycleEventHandler, LifeCycleManager}
+import wvlet.airframe.Design
+import wvlet.airframe.lifecycle.{LifeCycleEventHandler, LifeCycleManager}
 import wvlet.airspec.AirSpec
 
 /**
